Refcatored API calls
