+++ src/api/api.ts
@@ -1,0 +1,4 @@
+import account from "@/api/accountAPI";
+import school from "@/api/schoolAPI";
+
+export default { account: account, school: school }
