+++ src/store/accountStore.ts
@@ -1,6 +1,6 @@
 import Vue from 'vue'
 import Vuex from 'vuex'
-import api from "@/api"
+import api from "@/api/api"
 import {Account} from "@/models/Account";
 
 Vue.use(Vuex)
